+++ src/makegen.h
@@ -27,7 +27,7 @@
 #ifndef CWARE_MAKEGEN_H
 #define CWARE_MAKEGEN_H
 
-#define CWARE_MAKEGEN_VERSION   "1.0.0"
+#define CWARE_MAKEGEN_VERSION   "1.0.1"
 
 /* Common libraries */
 #include "carray.h"