+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGELOG.md
 
-DATE: June 6th, 2022
-VERSION: 1.0.0
+DATE: June 7th, 2022
+VERSION: 1.0.1
 
 # Features
-- Base program
+- Added header installation for libraries
+- Added uninstall target to library and project targets
+
+# Bug fixes
+- Headers are no longer executable (woop-de-doo)